Update types, files' paths
