--- native/src/navigation/OnboardingNavigator.tsx
@@ -1,21 +1,0 @@
-import * as React from "react";
-import { createStackNavigator } from "@react-navigation/stack";
-import { OnboardingStackParamList } from "./types/types";
-import {
-  SendPhoneNumberVerificationCode,
-  VerifyPhoneNumberCode,
-} from "../feature/onboarding/screen";
-
-const { Navigator, Screen } = createStackNavigator<OnboardingStackParamList>();
-
-const OnboardingNavigator = () => (
-  <Navigator>
-    <Screen
-      name="SendPhoneNumberVerificationCode"
-      component={SendPhoneNumberVerificationCode}
-    />
-    <Screen name="VerifyPhoneNumberCode" component={VerifyPhoneNumberCode} />
-  </Navigator>
-);
-
-export default OnboardingNavigator;
